refactor: change Integer to primitive types in response DTOs and repositories

+++ src/main/java/com/fitassist/backend/dto/response/activity/ActivityCalculatedResponseDto.java
@@ -27,7 +27,7 @@ public class ActivityCalculatedResponseDto implements Serializable {
 
 	private BigDecimal caloriesBurned;
 
-	private int time;
+	private Short time;
 
 	private BigDecimal weight;
